add method that alove to see average values

+++ src/main/java/Day.java
@@ -2,15 +2,15 @@
  * Created by Artem_Pryzhkov on 9/28/2016.
  */
 public class Day {
-    private double drunkWater;
+    private int drunkWater;
     private int eaten;
     private int walked;
 
-    public void drink(double waterLiters) {
+    public void drink(int waterLiters) {
         drunkWater+=waterLiters;
     }
 
-    public double getDrunk() {
+    public int getDrunk() {
         return drunkWater;
     }
